requirements.txt
flask==2.3.3
yt-dlp==2023.12.30
openai-whisper==20231117
openai==1.6.1
python-dotenv==1.0.0
requests==2.31.0